Refactoring the manifest to support all datacenters
There are three datacenters primarily used for testing. In order
to support all three environments, we need to add data center
location support in the manifest so that `cephci` can download
the necessary artifacts.

At the same time masking the datacenter location to the end user
i.e. the testers and CI executors.

Signed-off-by: Pragadeeswaran Sathyanarayanan <pragadeeswaran.sathyanarayanan@ibm.com>